Commonsguy pointed out that the "ClassNotFoundException when using custom Parcelable" (http://code.google.com/p/android/issues/detail?id=6822&can=1&q=Parcelable&sort=-stars&colspec=ID%20Type%20Status%20Owner%20Summary%20Stars)
issue applies to both Serializable and Parcelable so switching from one
to the other has not improved matters.

ajanhele...@gmail.com, commenting on the above issue states that: 
"As many of mentioned in various channels you can marshall/unmarshall
manually into a byte[] to work around this defect."

This is kind of awkward IMHO but I´ve implemented a SerializeHelper and
a SimpleSerializable interface to do just that with as little noise in
the calling code as possible.